Revert "clean up MI routines"
This reverts commit caeaa7c6f92d951fdbe9632bea1151e377a4bfc7.
