+++ configs/trg/MI_reg.yml
@@ -1,11 +1,11 @@
 # Mutual information regularization
 
-losstype: 'MINE_EMA'     # 'MINE' or 'MINE_EMA'
-alpha:  0.01             #  Exponential Moving Average (EMA) coupling parameter
-ma_eT:  1.0              #  EMA starting value
+losstype: 'MINE_EMA'        # 'MINE' or 'MINE_EMA'
+alpha:  0.01                #  Exponential Moving Average (EMA) coupling parameter
+ma_eT:  1.0                 #  EMA starting value
 
-beta:   -0.99             #  Positive for minimizing MI between (x,y), Negative for Maximizing (keep abs(beta) < 1)
-x_index: [0]             #  Variable indices (can contain several)
+beta:   0.99                #  Positive for minimizing MI between (x,y), Negative for Maximizing (keep abs(beta) < 1)
+x_index: [5]                #  Variable indices (can contain several)
 y_index: [1]
 
 hidden_dim: 64
